% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{makePopulationData}
\alias{makePopulationData}
\title{Generate population data from mrgsolve model with IIV}
\usage{
makePopulationData(model, n = 1000, seed = TRUE)
}
\arguments{
\item{model}{mrgsolve model object}

\item{n}{number of individuals to simulate}

\item{seed}{logical, if TRUE set a seed for reproducibility}
}
\value{
data.frame with n rows and columns for id and all parameters that are generated with IIV where applicable
}
\description{
decompose model to create a table of n indiviuals with all the necesary parameters to simulate a population profile
}
\examples{
\dontrun{
library(mrgsolve)
model <- mread("path_to_your_model.mod")
pop_data <- make_population_data(model, n = 500)
head(pop_data)
}

}
\author{
Romain Garreau
}
