% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{plotSim}
\alias{plotSim}
\title{Plot simulation results}
\usage{
plotSim(sim_data, unit = "mg/L", log_scale = FALSE)
}
\value{
plot of simulation results
}
\description{
Plot simulation results from mrgsolve simulation object
}
\examples{
\dontrun{
library(mrgsolve)
model <- mread("path_to_your_model.mod")
pop_data <- make_population_data(model, n = 500)
sim_data <- runSim(model, pop_data, amt = 1000)
plotSim(sim_data)
}
}
\author{
Romain Garreau
}
