% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{runSim}
\alias{runSim}
\title{Simulate population data using mrgsolve model}
\usage{
runSim(model, population_data, amt, plot = FALSE)
}
\arguments{
\item{model}{mrgsolve model object}

\item{population_data}{data.frame with individual parameters (output of makePopulationData)}

\item{amt}{numeric, amount to administer to each individual}

\item{plot}{logical, if TRUE return a plot of the simulation}
}
\value{
mrgsolve simulation object or plot if plot = TRUE
}
\description{
Simulate population data using a mrgsolve model and individual parameter data
}
\examples{
\dontrun{
library(mrgsolve)
model <- mread("path_to_your_model.mod")
pop_data <- make_population_data(model, n = 500)
sim_data <- runSim(model, pop_data, amt = 1000, plot = TRUE)
}
}
\author{
Romain Garreau
}
